\name{flow}
\alias{flow}
\alias{flow.default}
\alias{flow.htmlDocument}
\alias{flow.htmlElement}
\title{
  Layout HTML Content
}
\description{
  This function generates information about the layout of HTML content.
}
\usage{
flow(html, \dots)
\method{flow}{default}(html, \dots, css="", assets=NULL)
\method{flow}{htmlDocument}(html, width=NULL, height=NULL,
                            fonts="sans", device=currentDevice(),
                            engine=getOption("layoutEngine.backend"), \dots)
\method{flow}{htmlElement}(html, width=NULL, height=NULL,
                           fonts="sans", device=currentDevice(),
                           engine=getOption("layoutEngine.backend"), \dots)
}
\arguments{
  \item{html}{
    A character vector, an \code{"htmlDocument"} object
    (as generated by \code{\link{htmlDocument}}, or
    an \code{"htmlElement"} object (as generated by \code{\link{htmlElement}}). 
  }
  \item{css}{
    A character vector describing CSS styling for the HTML content.
  }
  \item{assets}{
    Zero or more paths to external assets (e.g., image files) for
    the HTML content.
  }
  \item{width, height}{
    \pkg{grid} units specifying the total width and height to use for the
    layout. 
  }
  \item{fonts}{
    A character vector containing R font family names.
  }
  \item{device}{
    The name of the device that will be used to render the layout.
  }
  \item{engine}{
    The backend that will perform the layout (e.g.,
    \code{cssboxEngine} from the \pkg{layoutEngineCSSBox} package).
  }
  \item{\dots}{
    Arguments used by methods.
  }
}
\value{
  A \code{"flowedhtml"} object.
}
\author{
  Paul Murrell
}
\seealso{
  \code{\link{grid.html}}
}
\examples{
flow("<p>test</p>")
}
\keyword{ dplot }
