\name{grid.html}
\alias{grid.html}
\alias{htmlGrob}
\alias{htmlGrob.default}
\alias{htmlGrob.htmlDocument}
\alias{htmlGrob.flowedhtml}
\title{
  Render HTML Content
}
\description{
  This function renders HTML content.
}
\usage{
grid.html(\dots)
htmlGrob(html, \dots)
\method{htmlGrob}{default}(html, \dots, css="", assets=NULL)
\method{htmlGrob}{htmlDocument}(html, 
                                x=0.5, y=0.5,
                                width=NULL, height=NULL,
                                default.units="npc",
                                just="centre",
                                fonts="sans",
                                device=currentDevice(),
                                engine=getOption("layoutEngine.backend"),
                                viewports=FALSE,
                                gp=gpar(), name=NULL, ...)
\method{htmlGrob}{flowedhtml}(html, 
                              x=0.5, y=0.5, 
                              default.units="npc",
                              just="centre",
                              viewports=FALSE,
                              gp=gpar(), name=NULL, ...) 
}
\arguments{
  \item{html}{
    A character vector, an \code{"htmlDocument"} object
    (as generated by \code{\link{htmlDocument}}, or
    an \code{"flowedhtml"} object (as generated by \code{\link{flow}}). 
  }
  \item{css}{
    A character vector describing CSS styling for the HTML content.
  }
  \item{assets}{
    Zero or more paths to external assets (e.g., image files) for
    the HTML content.
  }
  \item{x, y}{
    Numeric values or \pkg{grid} units specifying the location
    of the rendered output.
  }
  \item{width, height}{
    \pkg{grid} units specifying the total width and height to use for the
    layout. 
  }
  \item{default.units}{
    The \pkg{grid} unit used if the locations or dimensions are given
    as just numeric values.
  }
  \item{just}{
    Justification of the rendering relative to its (x, y) location.
  }
  \item{fonts}{
    A character vector containing R font family names.
  }
  \item{device}{
    The name of the device that will be used to render the layout.
  }
  \item{engine}{
    The backend that will perform the layout (e.g.,
    \code{cssboxEngine} from the \pkg{layoutEngineCSSBox} package).
  }
  \item{viewports}{
    A logical value indicating whether to generate \pkg{grid} viewports
    for each laid out HTML element.
  }
  \item{gp}{Graphical parameters, typically the output from a call
    to the function \code{gpar}.}
  \item{name}{A character identifier.}
  \item{\dots}{
    Arguments passed to \code{htmlGrob} or used by methods.
  }
}
\value{
  An \code{"htmlgrob"} object or a \code{"flowedhtmlgrob"} object.
}
\author{
  Paul Murrell
}
\seealso{
  \code{\link{flow}}
}
\examples{
grid.html("<p>test</p>")
}
\keyword{ aplot }
