\name{makeEngine}
\alias{makeEngine}
\alias{makeLayout}
\alias{copyAssets}
\alias{fontFiles}
\alias{layoutFields}
\title{
  Define a Backend 
}
\description{
  These functions can be used to define a backend that is used to
  perform HTML layout.
}
\usage{
makeEngine(layout, cssTransform=list())
makeLayout(type, name, x, y, width, height, baseline, text, family, 
    bold, italic, size, color, backgroundColor, borderLeftWidth, 
    borderTopWidth, borderRightWidth, borderBottomWidth, borderLeftStyle, 
    borderTopStyle, borderRightStyle, borderBottomStyle, borderLeftColor, 
    borderTopColor, borderRightColor, borderBottomColor)
copyAssets(html, dir) 
fontFiles(fonts, device)
layoutFields
}
\arguments{
  \item{layout}{
    A function that performs HTML layout and returns the result
    (which should be the result of calling \code{makeLayout}).
  }
  \item{cssTransform}{
    A list of (optional) functions that a backend
    can define for transforming CSS properties.  Currently used
    components are \code{fontWeight} and \code{fontFile}.
  }
  \item{type, name, x, y, width, height, baseline, text, family, 
    bold, italic, size, color, backgroundColor, borderLeftWidth, 
    borderTopWidth, borderRightWidth, borderBottomWidth, borderLeftStyle, 
    borderTopStyle, borderRightStyle, borderBottomStyle, borderLeftColor, 
    borderTopColor, borderRightColor, borderBottomColor}{
    Each argument represents a field that the layout result should
    return for each node in the HTML content.
    The type, name, location and size fields must be present, but
    most other fields can be \code{NA}.  In particular, the fields
    relating to text content should be \code{NA} for non-text nodes.
  }
  \item{html}{
    A \code{"flowedhtml"} object (as generated by the \code{\link{flow}}
    function.
  }
  \item{dir}{
    A directory to copy asset files to.
  }
  \item{fonts}{
    A character vector containing R font family names.
  }
  \item{device}{
    The name of the device that will be used to render the layout.
  }
}
\details{
  The \code{makeEngine} function should be used to create a backend
  for the \pkg{layoutEngine} package.  This function should be supplied
  a function that accepts arguments \code{html}, \code{width},
  \code{height}, \code{fonts}, and
  \code{device}, where \code{html} is an \code{"htmlDocument"} object
  (as produced by \code{\link{htmlDocument}} or
  \code{\link{htmlElement}}), \code{width} and \code{height} are
  the width and height that should be used to layout the HTML (in inches),
  and the other arguments are as described on this page.  The function
  should return a \code{"flowedhtml"} object (as produced by
  \code{"makeLayout"}).
  
  The remainder of the
  functions are helper functions for use in the main backend function:
  \code{copyAssets} to copy assets to a useful place; \code{fontFiles}
  to locate font files (probably to copy them as well); and
  \code{makeLayout} to create the result of the layout.
  The \code{layoutFields} object is useful for providing the names
  of the columns in the layout result.
}
\value{
  \code{makeEngine} produces a \code{"layoutengine"} object.
  
  \code{makeLayout} produces a \code{"flowedhtml"} object.  At present
  this is just a data frame, but longer term something more hierarchical
  is probably going to be better.
}
\author{
  Paul Murrell
}
\keyword{ dplot }

